+++ lib/src/service/image_service.g.dart
@@ -6,7 +6,7 @@ part of 'image_service.dart';
 // RetrofitGenerator
 // **************************************************************************
 
-// ignore_for_file: unnecessary_brace_in_string_interps
+// ignore_for_file: unnecessary_brace_in_string_interps,no_leading_underscores_for_local_identifiers
 
 class _ImageService implements ImageService {
   _ImageService(this._dio, {this.baseUrl});