+++ lib/src/service/live_input_service.g.dart
@@ -6,7 +6,7 @@ part of 'live_input_service.dart';
 // RetrofitGenerator
 // **************************************************************************
 
-// ignore_for_file: unnecessary_brace_in_string_interps
+// ignore_for_file: unnecessary_brace_in_string_interps,no_leading_underscores_for_local_identifiers
 
 class _LiveInputService implements LiveInputService {
   _LiveInputService(this._dio, {this.baseUrl});